+++ src/app/src/constants.ts
@@ -1,4 +1,5 @@
 import { AmountCategory } from './types';
+import { PathOptions } from 'leaflet';
 
 export const AMERICA_CENTER: [number, number] = [37.09024, -95.712891];
 
@@ -24,3 +25,16 @@ export const AMOUNT_CATEGORIES: AmountCategory[] = [
         size: 15,
     },
 ];
+
+export const STATE_STYLE_BASE = Object.freeze(
+    { color: 'black', weight: 0.62, fill: true, fillColor: 'white' }
+);
+
+export const STATE_STYLE_HOVER: PathOptions = Object.freeze(
+    {
+        fill: true,
+        fillColor: '#D0DAFF',
+        color: '#465EB5',
+        weight: 2,
+    }
+);